Move store to error_recovery.settings.
