+++ robot-server/robot_server/error_recovery/settings/store.py
